refact: test contact
