+++ tests/factories/contact-factory.ts
@@ -2,6 +2,34 @@ import { contact } from "@prisma/client";
 import { prisma } from "../../src/config";
 import { faker } from "@faker-js/faker";
 import { createUser } from "./users-factory";
+import { createAdmin } from "./admin-factory";
+import { generateValidAdminToken } from "../helpers";
+
+export function generateValidInsertInput() {
+  return {
+    name: faker.name.fullName(),
+    email: faker.internet.email(),
+    telephone: faker.phone.number("2299286####"),
+    message: faker.lorem.sentences(4),
+  };
+}
+
+export function generateValidDeleteInput() {
+  return {
+    id: faker.datatype.number(),
+  };
+}
+
+export async function newContact() {
+  const user = await createUser();
+  const admin = await createAdmin();
+  const token = await generateValidAdminToken(admin);
+  const input = generateValidInsertInput();
+
+  const contact = await createContact({ ...input, userId: user.id });
+
+  return { token, input, contact };
+}
 
 export async function createContact(data: Partial<contact> = {}): Promise<contact> {
   let userId: number;
